feat: Extend webpack plugin for supporting contenthash
