+++ src/alphaTab.worker.ts
@@ -1,2 +1,3 @@
+/**@target web */
 import * as alphaTab from './alphaTab.core';
 alphaTab.Environment.initializeWorker();
