+++ src/alphaTab.worklet.ts
@@ -1,2 +1,3 @@
+/**@target web */
 import * as alphaTab from './alphaTab.core';
 alphaTab.Environment.initializeAudioWorklet();
