+++ types/webpack/Identifier.d.ts
@@ -1,0 +1,5 @@
+declare module 'webpack/lib/util/identifier' {
+    export const contextify:  {
+        bindContextCache(context: string, associatedObjectForCache: any): (s: string) => string;
+    }
+}
