+++ types/webpack/MakeSerializable.d.ts
@@ -1,0 +1,4 @@
+declare module 'webpack/lib/util/makeSerializable' {
+    export default function (Constructor: any, request: string): void;
+    export default function (Constructor: any, request: string, name: string | null): void;
+}
