+++ types/webpack/index.d.ts
@@ -1,1 +1,3 @@
 import './ModuleTypeConstants'
+import './MakeSerializable'
+import './Identifier'
